Refactor dir structure for metric and tag registries
