+++ runner/src/coval_bench/runner/config.py
@@ -11,30 +11,9 @@ at deploy time without a code change.
 
 from __future__ import annotations
 
-from enum import StrEnum
-
 from pydantic import BaseModel, Field
 
-
-class TagCategory(StrEnum):
-    """Grouping for model tags, used to organize leaderboard filters."""
-
-    CAPABILITY = "capability"
-
-
-class ModelTag(StrEnum):
-    """Filterable/sortable model attributes surfaced on the leaderboard."""
-
-    REALTIME = "realtime"
-
-
-TAG_CATEGORIES: dict[ModelTag, TagCategory] = {
-    ModelTag.REALTIME: TagCategory.CAPABILITY,
-}
-
-if TAG_CATEGORIES.keys() != set(ModelTag):
-    _missing = ", ".join(sorted(set(ModelTag) - TAG_CATEGORIES.keys()))
-    raise RuntimeError(f"TAG_CATEGORIES is missing categories for: {_missing}")
+from coval_bench.registries import ModelTag
 
 
 class ProviderEntry(BaseModel):
